chore: renamed class
