+++ src/main/java/com/topy/bookreview/redis/RedisManager.java
@@ -1,12 +1,13 @@
 package com.topy.bookreview.redis;
 
+import java.util.concurrent.TimeUnit;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class RedisUtils {
+public class RedisManager {
 
   private final RedisTemplate<String, Object> redisTemplate;
 
@@ -14,8 +15,17 @@ public class RedisUtils {
     redisTemplate.opsForValue().set(key, value);
   }
 
+  public void save(String key, Object value, long expiryMillis) {
+    redisTemplate.opsForValue().set(key, value);
+    redisTemplate.expire(key, expiryMillis, TimeUnit.MILLISECONDS);
+  }
+
   public Object get(String key) {
     return redisTemplate.opsForValue().get(key);
+  }
+
+  public Long getExpire(String key, TimeUnit timeUnit) {
+    return redisTemplate.getExpire(key, timeUnit);
   }
 
   public void delete(String key) {
